add rsdGLClearErrors()
rsdGLClearErrors clears the gl error flag to ensure that
subsequent calls to glGetError reports the error flag of the last GL call,
not of the last failed GL call.

glGetError:
  When an error occurs, the error flag is set to the appropriate error code value
  No other errors are recorded until glGetError is called, the error code is
  returned, and the flag is reset to GL_NO_ERROR.

This was found by running RsHelloWorld.
In renderscript, glBindTexture reported a GL_INVALID_ENUM error,
while Mesa was ok.
Instead, the error flag was previously triggered by :
   GL_INVALID_ENUM in glGetFloatv(2pname=GL_MAX_TEXTURE_MAX_ANISOTROPY_EXT)
which wasn't checked for errors.

Change-Id: I88b9172935f6e22f0e98bd23ee654dc8a3d27532