Align renderscript Allocation memory for AVX
The Allocation class simply used malloc to allocate its memory, which
broke the use of the AVX vmovap* instructions (there may be SSE mova*
instructions in there too, I don't know) that require aligned memory.
Simply align it to 32 bytes always for now, as that works everywhere
and is only a minimal overhead (and even then only for trivially small
arrays).

Note that the C++ API (but not, I think the Java layer) appears to
allow for user-supplied pointers, so this problem may resurface
elsewhere.

Issue: AXIA-1222
Change-Id: I3dd1bc5eec4784d2ed97a8b903777ec791c5f913
Signed-off-by: Andy Ross <andy.ross@windriver.com>